+++ src/components/uploadsection.tsx
@@ -1,17 +1,43 @@
 import React, { useState } from 'react'
+import { useStores } from '@/core/stores/UseStores'
 import { UploadButton } from '@/utils/uploadthing'
 import '@uploadthing/react/styles.css'
 
 const UploadSection: React.FC = () => {
+  const { authStore } = useStores()
   const [hasError, setHasError] = useState<boolean>(false)
   const [errorMessage, setErrorMessage] = useState<string>('')
-
   const handleError = (error: string): void => {
     setErrorMessage(error)
     setHasError(true)
     setTimeout(() => {
       setHasError(false)
     }, 5000)
+  }
+
+  const businessid = authStore.userProfile?.profile._id
+  const data = {
+    businessId: businessid,
+    originalname: '',
+    blobname: '',
+    path: ''
+  }
+
+  const sendFileData = (): void => {
+    fetch('http://localhost:5000/api/file/fileupload', {
+      method: 'POST',
+      mode: 'cors',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(data)
+    })
+      .then(res => {
+        console.log(res)
+      })
+      .catch(err => {
+        throw err
+      })
   }
 
   return (
@@ -37,8 +63,15 @@ const UploadSection: React.FC = () => {
                   return 'Only .txt files allowed, file size up to 4MB'
                 }
               }}
-              onClientUploadComplete={res => {
-                console.log('Upload Complete')
+              onClientUploadComplete={(res) => {
+                console.log(res)
+                if (res != null) {
+                  const { key, url, name } = res[0]
+                  data.blobname = key
+                  data.originalname = name
+                  data.path = url
+                  sendFileData()
+                }
               }}
               onUploadError={(error: Error) => {
                 handleError(error.message)
